DL-3872 - Added Scala cross-compile
